XCOFF: have jump,make_fcontext show up as importable
On AIX, building with aix-soname=svr4, nm is used to list public symbols
in object files. This fixes the assembly to have make_fcontext and
jump_fcontext show up as public symbols.

The problem actually shows up when linking the coroutine library against
the context library when built with aix-soname=svr4.
However, except for this patch allowing to compile and link even
coroutine, I'm unsure if the assembly is correct from all aspects, like
alignment, TOC, and whether it works.

Thx to Michael Haubenwallner