Mark 32-bit GNU assembly files targeting Windows as /safeseh linking compatible.
We set the LSB of the magic symbol @feat.00 to 1.

This is used to communicate from the compiler to the linker, and specifically to express that the object file has opted into "safeseh"; any SEH handlers used in this file must be listed in the .sxdata section.

Since we don't have any SEH handlers in these files, this is trivially satisfied.

Reference: the PE-COFF specification, https://docs.microsoft.com/en-us/windows/win32/debug/pe-format#the-sxdata-section